+++ src/Mailables/AbstractMailable.php
@@ -21,17 +21,17 @@ abstract class AbstractMailable extends BaseMailable
     /**
      * @var string First line of the email body
      */
-    public $greeting = '';
+    public $greeting;
 
     /**
      * @var string Recipient email address
      */
-    public $email = '';
+    public $email;
 
     /**
      * @var string Subject of the email
      */
-    public $title = '';
+    public $title;
 
     /**
      * @var array|string Array of strings to be displayed as paragraphs
@@ -65,8 +65,8 @@ abstract class AbstractMailable extends BaseMailable
         $this->title = $title ?? $this->title;
         $this->messages = $messages ?? $this->messages;
         $this->call_to_action = $call_to_action ?? $this->call_to_action;
-        $this->view = $view ?? $this->view;
 
+        $this->view = $this->setView($view);
         $this->onQueue(config('post-office.queue'));
         $this->onConnection(config('post-office.driver'));
     }
@@ -76,7 +76,7 @@ abstract class AbstractMailable extends BaseMailable
      *
      * @return $this
      */
-    public function build()
+    public function build(): self
     {
         return $this->subject($this->title)
             ->view($this->view, [
@@ -86,5 +86,20 @@ abstract class AbstractMailable extends BaseMailable
                 'messages' => $this->messages,
                 'call_to_action' => $this->call_to_action,
             ]);
+    }
+
+    /**
+     * Set the Mailable's $view property.
+     *
+     *  1. $view parameter passed to the constructor
+     *  2. $view property declared in child `AbstractMailable` extension
+     *  3. 'post-office.mailable.view' config value (defaults to post-office::email)
+     *
+     * @param string|null $view
+     * @return string
+     */
+    private function setView(string $view = null): string
+    {
+        return $view ?? $this->view ?? config('post-office.mailables.view');
     }
 }